requirements.txt
# Core dependencies
python-dotenv>=0.21.0
psycopg2-binary>=2.9.0
requests>=2.28.0
numpy>=1.23.0
pandas>=1.5.0